requirements.txt
langchain>=0.0.267
openai>=1.0.0
faiss-cpu>=1.7.4
pandas>=1.5.3
numpy>=1.23.5
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
streamlit>=1.25.0